Refactor paths
