+++ proxy-supporter/src/api/rest/sitemaps/routes.js
